usb: usbpsu: Fixed MISRAC rule 10.6 violations
Fixed MISRA C-2012 Rule 10.6  below warning:
"misra_c_2012_rule_10_6_violation: Assigning expression
InstancePtr->Config.BaseAddress of width 32 to a target
of width 64."
Replaced u32 data type with UINTPTR.

misra_c_2012_rule_10_6_violation: Assigning composite expression
"(UsbEpNum << 1U) | Dir" of width 8 to a target of width 32.
Assign proper type cast to match assignement requirement.


Acked-for-series: Srinivas Goud <srinivas.goud@xilinx.com>

+++ XilinxProcessorIPLib/drivers/usbpsu/src/xusbpsu_command.c
@@ -94,7 +94,7 @@ s32 XUsbPsu_EpEnable(struct XUsbPsu *InstancePtr, u8 UsbEpNum, u8 Dir,
 					  (Dir == XUSBPSU_EP_DIR_OUT));
 	Xil_AssertNonvoid((Maxsize >= 64U) && (Maxsize <= 1024U));
 
-	PhyEpNum = XUSBPSU_PhysicalEp(UsbEpNum, Dir);
+	PhyEpNum = (u32)XUSBPSU_PhysicalEp(UsbEpNum, Dir);
 	Ept = &InstancePtr->eps[PhyEpNum];
 
 	Ept->UsbEpNum	= UsbEpNum;
@@ -186,7 +186,7 @@ s32 XUsbPsu_EpDisable(struct XUsbPsu *InstancePtr, u8 UsbEpNum, u8 Dir)
 	Xil_AssertNonvoid((Dir == XUSBPSU_EP_DIR_IN) ||
 						(Dir == XUSBPSU_EP_DIR_OUT));
 
-	PhyEpNum = XUSBPSU_PhysicalEp(UsbEpNum, Dir);
+	PhyEpNum = (u32)XUSBPSU_PhysicalEp(UsbEpNum, Dir);
 	Ept = &InstancePtr->eps[PhyEpNum];
 
 	/* make sure HW endpoint isn't stalled */
@@ -234,7 +234,7 @@ s32 XUsbPsu_SendEpCmd(struct XUsbPsu *InstancePtr, u8 UsbEpNum, u8 Dir,
 	Xil_AssertNonvoid((Dir == XUSBPSU_EP_DIR_IN) ||
 					  (Dir == XUSBPSU_EP_DIR_OUT));
 
-	PhyEpNum = XUSBPSU_PhysicalEp(UsbEpNum, Dir);
+	PhyEpNum = (u32)XUSBPSU_PhysicalEp(UsbEpNum, Dir);
 
 	XUsbPsu_WriteReg(InstancePtr, XUSBPSU_DEPCMDPAR0(PhyEpNum),
 					 Params->Param0);
